+++ src/vrm/spring_bone/update.rs
@@ -24,11 +24,15 @@ fn update_spring_bones(
     mut transforms: Query<(&mut Transform, &mut GlobalTransform)>,
     mut joints: Query<(&ChildOf, &mut SpringJointState, &SpringJointProps)>,
     spring_roots: Query<&SpringRoot>,
-    colliders: Query<&ColliderShape>,
     time: Res<Time>,
 ) {
     let delta_time = time.delta_secs();
     for spring_root in spring_roots.iter() {
+        let center_gtf = spring_root
+            .center_node
+            .and_then(|center| transforms.get(center).ok())
+            .map(|(_, gtf)| gtf)
+            .copied();
         for joint in spring_root.joints.iter().copied() {
             let Ok((child_of, mut state, props)) = joints.get_mut(joint) else {
                 continue;
@@ -43,7 +47,9 @@ fn update_spring_bones(
                 continue;
             };
 
-            let inertia = (state.current_tail - state.prev_tail) * (1. - props.drag_force);
+            let current_tail = center_local_to_global(state.current_tail, &center_gtf);
+            let prev_tail = center_local_to_global(state.prev_tail, &center_gtf);
+            let inertia = (current_tail - prev_tail) * (1. - props.drag_force);
             let stiffness = delta_time
                 * (parent_global_rotation
                     * state.initial_local_rotation
@@ -51,22 +57,21 @@ fn update_spring_bones(
                     * props.stiffness);
             let external = delta_time * props.gravity_dir * props.gravity_power;
 
-            let next_tail = state.current_tail + inertia + stiffness + external;
+            let next_tail = current_tail + inertia + stiffness + external;
             let mut next_tail =
                 head_global_pos + (next_tail - head_global_pos).normalize() * state.bone_length;
 
-            collision(
+            apply_collision(
                 &mut next_tail,
                 spring_root.colliders.iter().copied(),
                 props.hit_radius,
                 head_global_pos,
                 state.bone_length,
                 &transforms,
-                &colliders,
             );
 
             state.prev_tail = state.current_tail;
-            state.current_tail = next_tail;
+            state.current_tail = global_to_center_local(next_tail, &center_gtf);
 
             let to = (parent_gtf.compute_matrix() * state.initial_local_matrix)
                 .inverse()
@@ -84,23 +89,41 @@ fn update_spring_bones(
     }
 }
 
-fn collision(
+fn center_local_to_global(
+    tail_pos: Vec3,
+    center_gtf: &Option<GlobalTransform>,
+) -> Vec3 {
+    if let Some(gtf) = center_gtf.as_ref() {
+        gtf.transform_point(tail_pos)
+    } else {
+        tail_pos
+    }
+}
+
+fn global_to_center_local(
+    tail_pos: Vec3,
+    center_gtf: &Option<GlobalTransform>,
+) -> Vec3 {
+    if let Some(gtf) = center_gtf.as_ref() {
+        gtf.compute_matrix().inverse().transform_point3(tail_pos)
+    } else {
+        tail_pos
+    }
+}
+
+fn apply_collision(
     next_tail: &mut Vec3,
-    collider_entities: impl Iterator<Item = Entity>,
+    collider_entities: impl Iterator<Item = (Entity, ColliderShape)>,
     joint_radius: f32,
     head_global_pos: Vec3,
     bone_length: f32,
     transforms: &Query<(&mut Transform, &mut GlobalTransform)>,
-    colliders: &Query<&ColliderShape>,
 ) {
-    for collider in collider_entities {
-        let Ok(collider_shape) = colliders.get(collider) else {
-            continue;
-        };
+    for (collider, collider_shape) in collider_entities {
         let Ok((_, collider_gtf)) = transforms.get(collider) else {
             continue;
         };
-        collider_shape.calc_collision(
+        collider_shape.apply_collision(
             next_tail,
             collider_gtf,
             head_global_pos,